Added CatalogFragment.

+++ app/src/main/java/ru/klauz42/yetanotheronlinestore/domain/models/entities/Feedback.kt
@@ -2,5 +2,5 @@ package ru.klauz42.yetanotheronlinestore.domain.models.entities
 
 data class Feedback(
     val count: Int = 0,
-    val rating: Double = 0.0
+    val rating: Float = 0.0f
 )
